+++ src/macros.rs
@@ -1,10 +1,15 @@
 macro_rules! algos {
-    (@algo $algo:ident [$algo_s:expr] $decoder:ident $encoder:ident<$inner:ident> $({ $($constructor:tt)* })*) => {
+    (@algo $algo:ident [$algo_s:expr] $decoder:ident $encoder:ident<$inner:ident>
+        { @enc $($encoder_methods:tt)* }
+        { @dec $($decoder_methods:tt)* }
+    ) => {
         #[cfg(feature = $algo_s)]
         decoder! {
             #[doc = concat!("A ", $algo_s, " decoder, or decompressor")]
             #[cfg(feature = $algo_s)]
             $decoder
+
+            { $($decoder_methods)* }
         }
 
         #[cfg(feature = $algo_s)]
@@ -15,12 +20,15 @@ macro_rules! algos {
                 pub fn new(inner: $inner) -> Self {
                     Self::with_quality(inner, crate::Level::Default)
                 }
-            } $({ $($constructor)* })*
+            }
+
+            { $($encoder_methods)* }
         }
     };
 
     ($($mod:ident)::+<$inner:ident>) => {
-        algos!(@algo brotli ["brotli"] BrotliDecoder BrotliEncoder<$inner> {
+        algos!(@algo brotli ["brotli"] BrotliDecoder BrotliEncoder<$inner>
+        { @enc
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 let params = brotli::enc::backward_references::BrotliEncoderParams::default();
                 Self {
@@ -30,9 +38,13 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
 
-        algos!(@algo bzip2 ["bzip2"] BzDecoder BzEncoder<$inner> {
+        algos!(@algo bzip2 ["bzip2"] BzDecoder BzEncoder<$inner>
+        { @enc
+
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -41,9 +53,12 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
 
-        algos!(@algo deflate ["deflate"] DeflateDecoder DeflateEncoder<$inner> {
+        algos!(@algo deflate ["deflate"] DeflateDecoder DeflateEncoder<$inner>
+        { @enc
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -52,9 +67,13 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
 
-        algos!(@algo gzip ["gzip"] GzipDecoder GzipEncoder<$inner> {
+        algos!(@algo gzip ["gzip"] GzipDecoder GzipEncoder<$inner>
+        { @enc
+
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -63,9 +82,12 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
 
-        algos!(@algo zlib ["zlib"] ZlibDecoder ZlibEncoder<$inner> {
+        algos!(@algo zlib ["zlib"] ZlibDecoder ZlibEncoder<$inner>
+        { @enc
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -74,9 +96,13 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
 
-        algos!(@algo zstd ["zstd"] ZstdDecoder ZstdEncoder<$inner> {
+        algos!(@algo zstd ["zstd"] ZstdDecoder ZstdEncoder<$inner>
+        { @enc
+
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -106,9 +132,52 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
 
-        algos!(@algo xz ["xz"] XzDecoder XzEncoder<$inner> {
+            /// Creates a new encoder, using the specified compression level and pre-trained
+            /// dictionary, which will read uncompressed data from the given stream and emit a
+            /// compressed stream.
+            ///
+            /// Dictionaries provide better compression ratios for small files, but are required to
+            /// be present during decompression.
+            ///
+            /// # Errors
+            ///
+            /// Returns error when `dictionary` is not valid.
+            pub fn with_dict(inner: $inner, level: crate::Level, dictionary: &[u8]) -> ::std::io::Result<Self> {
+                Ok(Self {
+                    inner: crate::$($mod::)+generic::Encoder::new(
+                        inner,
+                        crate::codec::ZstdEncoder::new_with_dict(level.into_zstd(), dictionary)?,
+                    ),
+                })
+            }
+        }
+        { @dec
+            /// Creates a new decoder, using the specified compression level and pre-trained
+            /// dictionary, which will read compressed data from the given stream and emit an
+            /// uncompressed stream.
+            ///
+            /// Dictionaries provide better compression ratios for small files, but are required to
+            /// be present during decompression. The dictionary used must be the same as the one
+            /// used for compression.
+            ///
+            /// # Errors
+            ///
+            /// Returns error when `dictionary` is not valid.
+            pub fn with_dict(inner: $inner, dictionary: &[u8]) -> ::std::io::Result<Self> {
+                Ok(Self {
+                    inner: crate::$($mod::)+generic::Decoder::new(
+                        inner,
+                        crate::codec::ZstdDecoder::new_with_dict(dictionary)?,
+                    ),
+                })
+            }
+        }
+        );
+
+        algos!(@algo xz ["xz"] XzDecoder XzEncoder<$inner>
+        { @enc
+
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -117,9 +186,13 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
 
-        algos!(@algo lzma ["lzma"] LzmaDecoder LzmaEncoder<$inner> {
+        algos!(@algo lzma ["lzma"] LzmaDecoder LzmaEncoder<$inner>
+        { @enc
+
             pub fn with_quality(inner: $inner, level: crate::Level) -> Self {
                 Self {
                     inner: crate::$($mod::)+generic::Encoder::new(
@@ -128,6 +201,8 @@ macro_rules! algos {
                     ),
                 }
             }
-        });
+        }
+        { @dec }
+        );
     }
 }